🐛 fix(core): critical bug fixes and improved error logging
- Fix mask array assignment bug in frame_processor.py (blending was broken)
- Fix AMP dtype mismatch in losses.py (FP16 training was failing)
- Add NaN/Inf guard in dfl_module.py to prevent training corruption
- Add proper error logging to callbacks (preview/backup failures)
- Add resume state error logging in merger.py
- Add exception logging in frame_processor.py (3 locations)
- Fix FFmpeg stderr capture in video_io.py for debugging
- Improve GPEN fallback warning to guide users
- Enhance INT8 calibration warning with actionable guidance
- Fix missing 'Any' type import in tensorrt_runner.py

+++ visagen/training/losses.py
@@ -84,8 +84,8 @@ class DSSIMLoss(nn.Module):
         """
         c = pred.shape[1]
 
-        # Expand kernel for all channels
-        kernel = self.kernel.repeat(c, 1, 1, 1)
+        # Expand kernel for all channels (ensure dtype matches for AMP)
+        kernel = self.kernel.to(dtype=pred.dtype).repeat(c, 1, 1, 1)
 
         # Constants
         c1 = (self.k1 * self.max_val) ** 2
